+++ lib/core/design_system/constants/assets.dart
@@ -79,7 +79,7 @@ class Assets {
   static const String discountTag = '$iconsPath/discount-tag-02.svg';
   static const String packageAdd = '$iconsPath/package-add.svg';
   static const String cartPackageDelivered = '$iconsPath/package-delivered.svg';
-  // Account Screen
+  // account Screen
   static const String moreHorizontalSquare = '$iconsPath/ic_more_horizontal_square.svg';
   static const String package= '$iconsPath/ic_package.svg';
   static const String heartCheck= '$iconsPath/ic_heart_check.svg';
